backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
openpyxl==3.1.5
python-multipart==0.0.18
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
rapidfuzz==3.10.1
python-telegram-bot==21.6
